+++ src/protocol/types/responses/authorize.ts
